refactor: extract shared SVG attr parser; add changeset
Extract duplicated parseAttrs + num helpers used by both the geometry and
text-layout extractors into a single parseSvgAttrs module. Rename the local
Rect type in extractTextLayout to Bounds to avoid shadowing Figma's global
Rect. Document the spread-override pattern that lets extractTextLayout
override transformDimension/transformRotationAndPosition results.

Add changeset describing the editable ShapeWithText export.
