+++ .changeset/editable-shape-with-text.md
@@ -1,0 +1,5 @@
+---
+'penpot-exporter': minor
+---
+
+Export `ShapeWithText` nodes (FigJam / Figma Slides primitives like arrows, triangles, speech bubbles, stars) as an editable Penpot group containing an editable path and an editable text label, instead of a rasterized PNG.
