+++ plugin-src/translators/shapeWithText/index.ts
@@ -1,6 +1,7 @@
 export * from './exportShapeWithTextSvg';
 export * from './extractTextLayout';
 export * from './normalizeCommands';
+export * from './parseSvgAttrs';
 export * from './parseSvgTransform';
 export * from './translateShapeWithTextContent';
 export * from './translateShapeWithTextGeometry';
